game pixel size computed in function instead of stored in game_state, try to factor in menu height but not working yet

+++ game/include/render.h
@@ -48,7 +48,7 @@ extern Texture *empty_texture;
 
 void shader_set_texture(GLuint shader_id, Texture* texture);
 void shader_set_color(GLuint shader_id, Color color);
-void shader_set_transform_pixels(GLuint shader_id, u32 width, u32 height);
+void shader_set_transform_pixels(GLuint shader_id, f32 width, f32 height);
 
 Texture *create_texture(void* image_data, u32 width, u32 height);
 Texture *load_texture(const char* filename);
